v3/requirements.txt
# Core LLM
google-generativeai>=0.8.0
langchain-google-genai>=2.0.0

# Agent Orchestration (L5)
langgraph>=0.2.0

# MCP Protocol
mcp>=1.0.0

# RAG System (L3)
llama-index>=0.11.0
llama-index-vector-stores-chroma>=0.2.0
llama-index-embeddings-huggingface>=0.3.0
llama-index-llms-gemini>=0.3.0
chromadb>=0.5.0

# Web Search (L4)
tavily-python>=0.4.0

# Code Execution Backup
e2b-code-interpreter>=0.0.10

# Benchmarks
datasets>=2.20.0
Pillow>=10.0.0

# Scientific Computing
numpy>=1.26.0
scipy>=1.11.0
matplotlib>=3.8.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Utilities
httpx>=0.27.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.0.0
